requirements.txt
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0
SQLAlchemy==2.0.36
pyspellchecker==0.8.2